Fixes #165: Reduce initial size for metadata region (#167)
Introduced a macro MAX_DATAITEM_NUM_HINT which is a hint that a maximum of 1 million data items may be created. Multiplying that by the size of data item, we will set the initial size. It is then compared to 1/4th of the region size, using a conditional operator and the minimum of the two is selected. It is again compared to the MIN_HEAP_SIZE and the larger of the two is set as the initial size for metadata region.