Account for behavior change to CXI counters (#295)
Previous to NETCASSINI-6335, fi_cntr_read/readerr returned the
last cached value while dispatching a counter read
asynchronously. Some applications required that API adhere to
the defined semantic of returning the current value, which
required waiting for the counter to be returned and
NETCASSINI-6335 made this the default behavior. This impacted
the minimum latency of fam_quiet().

NETCASSINI-6335 additionally added the FI_CXI_CNTR_CACHED flag
to provide the previous behavior and this patch enables it, if it
is available in the installed libfabric.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>